Create useSuggestions hook with all online/offline logic for reuse in Match screen
