+++ src/components/Suggestions/SuggestionsContainer.tsx
@@ -15,27 +15,20 @@ import { log } from "sharedHelpers/logger";
 import {
   useLastScreen,
   useLocationPermission,
-  usePerformance
+  usePerformance,
+  useSuggestions
 } from "sharedHooks";
 import { isDebugMode } from "sharedHooks/useDebugMode";
 import useStore from "stores/useStore";
 
 import fetchCoarseUserLocation from "../../sharedHelpers/fetchUserLocation";
 import flattenUploadParams from "./helpers/flattenUploadParams";
-import isolateHumans, { humanFilter } from "./helpers/isolateHumans";
-import sortSuggestions from "./helpers/sortSuggestions";
 import useClearComputerVisionDirectory from "./hooks/useClearComputerVisionDirectory";
 import useNavigateWithTaxonSelected from "./hooks/useNavigateWithTaxonSelected";
-import useOfflineSuggestions from "./hooks/useOfflineSuggestions";
-import useOnlineSuggestions from "./hooks/useOnlineSuggestions";
 import Suggestions from "./Suggestions";
 import TaxonSearchButton from "./TaxonSearchButton";
 
 const logger = log.extend( "SuggestionsContainer" );
-
-const ONLINE_THRESHOLD = 78;
-// note: offline threshold may need to change based on input from the CV team
-const OFFLINE_THRESHOLD = 0.78;
 
 const setQueryKey = ( selectedPhotoUri, shouldUseEvidenceLocation ) => [
   "scoreImage",
@@ -180,17 +173,26 @@ const SuggestionsContainer = ( ) => {
   const shouldFetchOnlineSuggestions = ( hasPermissions !== undefined )
     && fetchStatus === "loading";
 
+  const onlineSuggestionsAttempted = fetchStatus === "online-fetched"
+    || fetchStatus === "online-error";
+
   const {
-    dataUpdatedAt: onlineSuggestionsUpdatedAt,
-    error: onlineSuggestionsError,
-    onlineSuggestions,
     timedOut,
-    resetTimeout
-  } = useOnlineSuggestions( {
+    resetTimeout,
+    onlineSuggestions,
+    offlineSuggestions,
+    onlineSuggestionsError,
+    onlineSuggestionsUpdatedAt,
+    suggestions,
+    usingOfflineSuggestions,
+    urlWillCrashOffline
+  } = useSuggestions( {
+    shouldFetchOnlineSuggestions,
     dispatch,
     flattenedUploadParams,
     queryKey,
-    shouldFetchOnlineSuggestions
+    selectedPhotoUri,
+    onlineSuggestionsAttempted
   } );
 
   const createUploadParams = useCallback( async ( uri, showLocation ) => {
@@ -203,33 +205,6 @@ const SuggestionsContainer = ( ) => {
   }, [
     currentObservation
   ] );
-
-  // 20240815 amanda - it's conceivable that we would want to use a cached image here eventually,
-  // since the user can see the small square version of this image in MyObs/ObsDetails already
-  // but for now, passing in an https photo to predictImage while offline crashes the app
-  const urlWillCrashOffline = selectedPhotoUri.includes( "https://" ) && !isConnected;
-
-  // skip to offline suggestions if internet connection is spotty
-  const tryOfflineSuggestions = !urlWillCrashOffline && (
-    timedOut
-    || (
-      ( !onlineSuggestions || onlineSuggestions?.results?.length === 0 )
-      && ( fetchStatus === "online-fetched" || fetchStatus === "online-error" ) )
-  );
-
-  const {
-    offlineSuggestions
-  } = useOfflineSuggestions( selectedPhotoUri, {
-    dispatch,
-    latitude: flattenedUploadParams?.lat,
-    longitude: flattenedUploadParams?.lng,
-    tryOfflineSuggestions
-  } );
-
-  const usingOfflineSuggestions = tryOfflineSuggestions || (
-    offlineSuggestions.length > 0
-      && ( !onlineSuggestions || onlineSuggestions?.results?.length === 0 )
-  );
 
   const setSelectedTaxon = taxon => {
     dispatch( {
@@ -267,10 +242,6 @@ const SuggestionsContainer = ( ) => {
     ]
   );
 
-  const unfilteredSuggestions = onlineSuggestions?.results?.length > 0
-    ? onlineSuggestions.results
-    : offlineSuggestions;
-
   const isLoading = fetchStatus === "loading";
 
   const { loadTime } = usePerformance( {
@@ -279,93 +250,6 @@ const SuggestionsContainer = ( ) => {
   if ( isDebugMode( ) && loadTime ) {
     logger.info( loadTime );
   }
-
-  const filterSuggestions = useCallback( ( suggestionsToFilter: Suggestion[] ) => {
-    const sortedSuggestions = sortSuggestions(
-      isolateHumans( suggestionsToFilter ),
-      { usingOfflineSuggestions }
-    );
-    const newSuggestions = {
-      ...initialSuggestions,
-      otherSuggestions: sortedSuggestions
-    };
-    // no suggestions
-    if ( sortedSuggestions.length === 0 ) {
-      return {
-        ...newSuggestions,
-        otherSuggestions: [],
-        topSuggestionType: "none"
-      };
-    }
-    // human top suggestion
-    if ( sortedSuggestions.find( humanFilter ) ) {
-      return {
-        ...newSuggestions,
-        topSuggestion: sortedSuggestions[0],
-        topSuggestionType: "human",
-        otherSuggestions: []
-      };
-    }
-
-    // Note: score_vision responses have combined_score values between 0 and
-    // 100, compared with offline model results that have scores between 0
-    // and 1
-    const filterCriteria = usingOfflineSuggestions
-      ? s => s.score > OFFLINE_THRESHOLD
-      : s => s.combined_score > ONLINE_THRESHOLD;
-
-    const suggestionAboveThreshold = _.find(
-      sortedSuggestions,
-      filterCriteria
-    );
-
-    if ( suggestionAboveThreshold ) {
-      // make sure we're not returning the top suggestion in Other Suggestions
-      const firstSuggestion = _.remove(
-        sortedSuggestions,
-        ( s: Suggestion ) => s.taxon.id === suggestionAboveThreshold.taxon.id
-      ).at( 0 );
-      return {
-        ...newSuggestions,
-        topSuggestion: firstSuggestion,
-        topSuggestionType: usingOfflineSuggestions
-          ? "above-offline-threshold"
-          : "above-online-threshold"
-      };
-    }
-    if ( !suggestionAboveThreshold && usingOfflineSuggestions ) {
-      // no top suggestion for offline
-      return {
-        ...newSuggestions,
-        topSuggestion: null,
-        topSuggestionType: "not-confident"
-      };
-    }
-
-    // online common ancestor
-    if ( onlineSuggestions?.common_ancestor ) {
-      return {
-        ...newSuggestions,
-        topSuggestion: onlineSuggestions?.common_ancestor,
-        topSuggestionType: "common-ancestor"
-      };
-    }
-
-    // no top suggestion
-    return {
-      ...newSuggestions,
-      topSuggestionType: "not-confident"
-    };
-  }, [
-    onlineSuggestions?.common_ancestor,
-    usingOfflineSuggestions
-  ] );
-
-  // since we can calculate this, there's no need to store it in state
-  const suggestions = useMemo(
-    ( ) => filterSuggestions( unfilteredSuggestions ),
-    [unfilteredSuggestions, filterSuggestions]
-  );
 
   const toggleLocation = useCallback( async ( { showLocation } ) => {
     const newImageParams = await createUploadParams( selectedPhotoUri, showLocation );
