+++ src/sharedHooks/index.js
@@ -26,6 +26,7 @@ export { default as useQuery } from "./useQuery";
 export { default as useRemoteObservation } from "./useRemoteObservation";
 export { default as useShare } from "./useShare";
 export { default as useStoredLayout } from "./useStoredLayout";
+export { default as useSuggestions } from "./useSuggestions/useSuggestions";
 export { default as useTaxon } from "./useTaxon";
 export { default as useTaxonSearch } from "./useTaxonSearch";
 export { default as useTranslation } from "./useTranslation";
