+++ src/sharedHooks/useSuggestions/useOfflineSuggestions.ts
